// packages/front/ia-dashboard/src/store/building/types.ts
export type BuildingType = {
  id: string | null;
  name: string | null;
  pref_treatment: string | null; // был boolean, но кто то проебался, и теперь это селект
  region: string | null;
  site_format: string[] | null;
  site_type: string[] | null;
  transaction_form: string[] | null;
  min_cost_object: string | null;
  max_cost_object: string | null;
  water_supply_available: boolean | null;
  min_water_supply_rate_consumption: string | null;
  max_water_supply_rate_consumption: string | null;
  min_water_supply_rate_transport: string | null;
  max_water_supply_rate_transport: string | null;
  min_water_supply_objects_max_capacity: string | null;
  max_water_supply_objects_max_capacity: string | null;
  gas_supply_available: true | null;
  gas_supply_rate_consumption_min: string | null;
  gas_supply_rate_consumption_max: string | null;
  min_gas_supply_rate_transport: string | null;
  max_gas_supply_rate_transport: string | null;
  electricity_supply_available: boolean | null;
  min_electricity_rate_consumption: string | null;
  max_electricity_rate_consumption: string | null;
  min_electricity_rate_transport: string | null;
  max_electricity_rate_transport: string | null;
  heating_available: boolean | null;
  min_heating_rate_consumption: string | null;
  max_heating_rate_consumption: string | null;
  min_heating_rate_transport: string | null;
  max_heating_rate_transport: string | null;
  waste_disposal_available: boolean | null;
  min_waste_disposal_rate: string | null;
  max_waste_disposal_rate: string | null;
  access_roads_available: boolean | null;
  railways_available: boolean | null;
  truck_parking_available: boolean | null;
  additional_comments: string | null;
  user: number | null;
  state: number | null;
};